// src/components/footer/enum.ts
enum TagNames {
  DIV = 'div',
  FOOTER = 'footer',
  A = 'a',
}

enum Styles {
  FOOTER = 'footer-new',
  CONTAINER = 'container-footer',
  LINK = 'link-footer',
}

enum Content {
  LINK_ABOUT_US = 'About us',
}

enum Attributes {
  HREF = 'href',
  HREF_VALUE_ABOUT_US = '/about_us',
}

export { TagNames, Styles, Content, Attributes };